+++ pandas_dataclasses/__init__.py
@@ -15,7 +15,7 @@ __all__ = [
 
 
 from . import core
-from .core.asdata import *
+from .core.aspandas import *
 from .core.mixins import *
 from .core.specs import *
 from .core.typing import *
